Chat IA ERP/python/requirements.txt
google-generativeai>=0.3.0
psycopg2-binary>=2.9.11
python-dotenv>=1.0.0

